backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.2.1
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
pymongo==4.13.0
pytz==2025.2
sqlparse==0.5.3
